backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
langchain==0.1.0
langgraph==0.0.20
langchain-openai==0.0.5
openai>=1.10.0
langchain-google-genai==0.0.6
google-generativeai==0.3.2
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
requests>=2.31.0